API calls now include / exclude editted pokemons
Old and new Pokemon data are now stored in local storage.
Before each request, the app checks if any editted Pokémon match the current filters (based on old values) — if so, they are excluded from the request.
If the filters match the new values, their IDs are included in the request.